arch-riscv: allow VLEN up to 512Ki bits and vl up to 2**17
Three limits stood between the offload path and a vector unit with
131072 32-bit elements (VLEN=524288 at LMUL=8):

- ExtMachInst.vl was Bitfield<57,41>, 17 bits, so it topped out at
  131071 - one short. Bits 58-59 were unused, so it widens to
  <58,41> without disturbing any other field.
- MaxVecLenInBits and RiscvVectorLength.max were both 65536. Raising
  them costs memory in gem5's own physical vector register file even
  when vector state lives in the offloaded unit, which is why the
  ceiling is 524288 rather than something larger.
- VecOffloadRecord.vl/vstart were uint16_t, so 131072 truncated to
  exactly 0 and every record crossed the bridge claiming an empty
  vector - a silent wrong answer rather than an error.

+++ src/arch/riscv/insts/vec_offload.hh
@@ -42,8 +42,8 @@ struct VecOffloadRecord
     uint8_t vlmul = 0;       // vtype encoding (3 bits, fractional incl.)
     uint8_t vta = 0;
     uint8_t vma = 0;
-    uint16_t vl = 0;
-    uint16_t vstart = 0;     // phase 1: always 0, still carried
+    uint32_t vl = 0;         // up to 2**17 elements (LMUL=8, VLEN=512Ki)
+    uint32_t vstart = 0;     // phase 1: always 0, still carried
     // scalar operand value (.vx/.vf forms), read at execute
     uint64_t scalar = 0;
 };
